Optimized many shaders

+++ Debug/shaders/stdlight.glsl
@@ -7,10 +7,10 @@
 //#include "playerlight.glsl"
 
 // shadow is smallest between shadow-value and daylight level
-float shadow = min(1.0, daylight * lightdata.x + brightness);
+lowp float shadow = min(1.0, daylight * lightdata.x + brightness);
 
 // corner shadows and dotlight
-float cornershad = smoothstep(0.2, 0.7, lightdata.z);
+lowp float cornershad = smoothstep(0.2, 0.7, lightdata.z);
 cornershad = min(1.0, cornershad * worldLight + brightness * 0.125);
 
 // scaled shadow color
